+++ 055_r_array_analysis.r
@@ -1140,7 +1140,44 @@ warning("Code construction in progress.")
 
 # Screen known obesity genes ----
 
-# As per 10.3390/ijms231911005 using 
+# Could be used to warrent DEG search despite PCA signal, as per 10.3390/ijms231911005 using:
+# Leptin (LEP), the leptin receptor (LEPR), proopiomelanocortin (POMC), prohormone convertase 1 (PCSK1), the melanocortin 4 receptor (MC4R), single-minded homolog 1 (SIM1), brain-derived neurotrophic factor (BDNF), and the neurotrophic tyrosine kinase receptor type 2 gene (NTRK2) 
+
+# 1.) Compile data ----
+
+SE_list <- list(BRAT, IWAT, LIVT, EVAT)
+names(SE_list) <- c("BRAT", "IWAT", "LIVT", "EVAT")
+obesity_genes <- c("LEP", "LEPR", "POMC", "PCSK1", "MC4R", "SIM1", "BDNF", "NTRK2")
+
+# 2.) Filter data to obesity genes and check ----
+
+SE_list_og <- lapply(SE_list, function(seob) seob[   which( toupper(rowData(seob)[["SYMBOL"]]) %in% obesity_genes ) ] )
+lapply(SE_list_og,  function(seob) toupper(rowData(seob)[["SYMBOL"]]))
+
+library(sechm) 
+
+# https://github.com/plger/sechm
+# https://www.bioconductor.org/packages/release/bioc/vignettes/sechm/inst/doc/sechm.html
+
+sechm(
+  SE_list_og[[1]],
+  top_annotation = "ParentalDietMoFa",
+  features = rownames(SE_list_og[[1]]),
+  do.scale = TRUE,
+  gaps_at = "ParentalDietMoFa",
+  right_annotation = "SYMBOL",
+  mark = c("Bdnf", "Lepr", "Lep", "Sim1", "Pomc", "Ntrk2", "Pcsk1", "Mc4r"),
+  show_rownames = FALSE)
+
+SE_list[[1]]
+
+
+# 2.) Filter data to only contain obesity genes ----
+
+SE_list[[1]][1:5,]
+
+rowRanges(SE_list[[1]]), 
+
 
 # Find Deferentially expressed genes ----